# Conflicts:
#	src/main/resources/item_photos/widelec.png

+++ src/main/java/po/grupa2/ksiazeczkagot/dto/Wspolrzedna.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class Wspolrzedna extends BaseDto<Long> {
   private int stopnie;
   private int minuty;
-  private int sekundy;
+  private double sekundy;
 }
